+++ apps/backoffice/src/app/(payload)/[[...segments]]/page.tsx
@@ -2,7 +2,7 @@
 /* DO NOT MODIFY IT BECAUSE IT COULD BE REWRITTEN AT ANY TIME. */
 import type { Metadata } from 'next';
 
-import config from '@/payload.config';
+import config from '@/payload-config';
 import { generatePageMetadata, RootPage } from '@payloadcms/next/views';
 
 import { importMap } from '../importMap';
