+++ apps/backoffice/src/app/(payload)/api/graphql/route.ts
@@ -1,6 +1,6 @@
 /* THIS FILE WAS GENERATED AUTOMATICALLY BY PAYLOAD. */
 /* DO NOT MODIFY IT BECAUSE IT COULD BE REWRITTEN AT ANY TIME. */
-import config from '@/payload.config';
+import config from '@/payload-config';
 import { GRAPHQL_POST, REST_OPTIONS } from '@payloadcms/next/routes';
 
 export const POST = GRAPHQL_POST(config);
